+++ modules/core/arrow-core-data/src/main/kotlin/arrow/core/Ior.kt
@@ -1,15 +1,6 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind
-import arrow.core.Either
-import arrow.core.Eval
-import arrow.core.Left
-import arrow.core.None
-import arrow.core.Option
-import arrow.core.Right
-import arrow.core.Some
-import arrow.core.Tuple2
-import arrow.core.identity
 import arrow.higherkind
 import arrow.typeclasses.Applicative
 import arrow.typeclasses.Semigroup
