+++ modules/core/arrow-core-data/src/test/kotlin/arrow/core/IorTest.kt
@@ -1,19 +1,10 @@
-package arrow.data
+package arrow.core
 
 import arrow.Kind2
-import arrow.core.Either
-import arrow.core.None
-import arrow.core.Some
 import arrow.core.extensions.eq
 import arrow.core.extensions.hash
 import arrow.core.extensions.semigroup
-import arrow.data.extensions.ior.applicative.applicative
-import arrow.data.extensions.ior.bifunctor.bifunctor
-import arrow.data.extensions.ior.eq.eq
-import arrow.data.extensions.ior.hash.hash
-import arrow.data.extensions.ior.monad.monad
-import arrow.data.extensions.ior.show.show
-import arrow.data.extensions.ior.traverse.traverse
+import arrow.data.*
 import arrow.data.Ior.Right
 import arrow.test.UnitSpec
 import arrow.test.laws.BifunctorLaws
@@ -21,13 +12,11 @@ import arrow.test.laws.HashLaws
 import arrow.test.laws.MonadLaws
 import arrow.test.laws.ShowLaws
 import arrow.test.laws.TraverseLaws
-import arrow.test.laws.fix
 import arrow.typeclasses.Eq
 import arrow.typeclasses.Hash
 import arrow.typeclasses.Monad
 import io.kotlintest.properties.forAll
 import io.kotlintest.runner.junit4.KotlinTestRunner
-import io.kotlintest.shouldBe
 import org.junit.runner.RunWith
 
 @RunWith(KotlinTestRunner::class)
